fix: return the requested sample count, and validate before sampling
evaluate_range returned a single sample whenever start_time equalled
end_time, no matter what count was asked for, which silently broke any
caller sizing a buffer from the count it passed in. The step is zero for
an empty range, not undefined, so the general path already handles it;
only a count of one needs a special case, to avoid dividing by zero.

The range was also validated after the early return for counts of one or
less, so evaluate_range(10, 0, 0) returned a value while the same
reversed range with a count of 2 threw. The range is now checked first,
for every count. A count of zero returns nothing rather than one sample,
and a negative count is rejected rather than quietly treated as one.

Animation::num_samples returns size_t, matching Channel::num_samples.
Both count the same thing and had no reason to differ.

+++ include/anim/animation.hpp
@@ -163,12 +163,13 @@ public:
      * The span is half-open, matching Channel::evaluate_range_by_rate(): the
      * result is length() * sample_rate rounded up, so 4 seconds at 30 Hz gives
      * 120 rather than 121. An animation with no channels gives 0; one with no
-     * length gives 1.
+     * length gives 1. Matches Channel::num_samples() in both convention and
+     * return type.
      * @param sample_rate Samples per unit time; must be positive.
      * @return Sample count (0 when there are no channels).
      * @throws std::invalid_argument if @p sample_rate is not positive.
      */
-    int num_samples(double sample_rate) const;
+    size_t num_samples(double sample_rate) const;
 
     /// @brief Equality across name, time range and channels.
     bool operator==(const Animation& other) const;
